cleanup and move input component to their own file
